pyproject.toml
[project]
name = "mevzuat-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "charset-normalizer>=3.4.4",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "langchain-core>=1.2.5",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "lxml>=6.0.2",
    "pgvector>=0.4.2",
    "psycopg-binary>=3.3.2",
    "psycopg>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
]
